refactor: make lib top-level
